now that Mask::count is unsigned, pow(double, uint) is ambiguous: explicit cast to int helps

+++ tests/arithmetics.cpp
@@ -376,7 +376,7 @@ template<typename Vec> void testProduct()
                     x2 *= x;
                 }
             } else {
-                x2 = static_cast<T>(pow(static_cast<double>(x), m.count()));
+                x2 = static_cast<T>(pow(static_cast<double>(x), static_cast<int>(m.count())));
             }
             COMPARE(v.product(m), x2) << m << v;
         } while (true);
